USB: core: add a way to revoke access to open USB devices
There is still a need for applications in user-space to open USB devices
directly, usually through libusb or bindings for it, to support all the
devices that don't have kernel-level class driver[1] and thus need to be
implemented in user-space.

Unfortunately, because of the way USB device nodes and devices are exposed
to user-space, we don't have any easy ways to revoke access to those
devices once the device node has been opened. It is potentially possible
for a user to open the device, and hold onto it even after a fast-user
switch, as if they were still in control of the device, even though
they wouldn't be allowed to open the device node again.

This commit implements the APIs used to revoke access to USB devices given:
- a file descriptor for a device node (with an optional UID check)
- or a file descriptor for a user namespace
- or a uid

If a file descriptor to a user namespace is provided, the uid is mapped
using that user namespace.

Access revocation is preferred to closing the file descriptor because it
avoids waking up all applications at once, which is especially important
when trying to start a new session when doing fast-user switching, and
also stops the application from trying to reopen the device and not
being able to show any messages to the user.

The input sub-system implements a similar API through an ioctl in
c7dc657 ("Input: evdev - add EVIOCREVOKE ioctl").

This commit does not include any user-space facing APIs.

[1]:
Non exhaustive list of devices and device types that need raw USB access:
- all manners of single-board computers and programmable chips and
devices (avrdude, STLink, sunxi bootloader, flashrom, etc.)
- 3D printers
- scanners
- LCD "displays"
- user-space webcam and still cameras
- game controllers
- video/audio capture devices
- sensors
- software-defined radios
- DJ/music equipment
- protocol analysers
- Rio 500 music player

Signed-off-by: Bastien Nocera <hadess@hadess.net>