usb: Implement usb_revoke_device() BPF function
This functionality allows a sufficiently privileged user-space process
to upload a BPF programme that will call to usb_revoke_device() as if
it were a kernel API.

Using BPF to implement this functionality means that we can use the same
code path to handle a device fd being passed to revoke access, or all the
devices a user has open, or devices for a specific user namespace.

This functionality will be used by logind to revoke access to devices on
fast user-switching to start with. Eventually application containers
would use this functionality to offer fine-grained access to USB
devices.

An test application using this API is available at:
https://github.com/hadess/usb-revoke-userspace/

Signed-off-by: Bastien Nocera <hadess@hadess.net>