HID: usbhid: Allow drivers to set the serial
Don't overwrite the "uniq" identifier that drivers could have set with
the USB serial. If there is a protocol built on top of HID, as done for
Logitech HID++ hardware, we want the specialised driver to set the serial
rather than the transport.

The USB serial is still available in the serial sysfs attribute of the
parent device in the USB subsystem.

Signed-off-by: Bastien Nocera <hadess@hadess.net>